+++ dbrequests/mysql/connection.py
@@ -4,13 +4,13 @@ compliant.
 """
 
 import logging
+import re
 from contextlib import contextmanager
 from inspect import getfullargspec as getargs
-from tempfile import NamedTemporaryFile as TmpFile
 
 from datatable import Frame, dt, f, join, rbind, str64
-
 from dbrequests import Connection as SuperConnection
+from dbrequests.temp_file import temp_file
 
 
 class Connection(SuperConnection):
@@ -88,13 +88,13 @@ class Connection(SuperConnection):
 
     def _send_data_insert(self, df, table):
         logging.info(f"sending data with insert: {df.shape[0]} rows")
-        with TmpFile(mode="w", newline="") as tf:
+        with temp_file() as tf:
             self._write_csv(df, tf)
             self._infile_csv(tf, df, table)
 
     def _send_data_replace(self, df, table):
         logging.info(f"sending data with replace: {df.shape[0]} rows")
-        with TmpFile(mode="w", newline="") as tf:
+        with temp_file() as tf:
             self._write_csv(df, tf)
             self._infile_csv(tf, df, table, replace="replace")
 
@@ -155,11 +155,15 @@ class Connection(SuperConnection):
             return None
         df = df[:, f[:].remove(f[:]).extend(str64(f[:]))][:, df.names]
         df.replace(None, "NULL")
-        df.to_csv(path=file.name, header=False)
+        df.to_csv(path=file, header=False)
 
     def _infile_csv(self, file, df, table, replace=""):
-        self.bulk_query(
-            """
+        # On Windows paths are denoted by '\\'. A backslash in the sql statement
+        # has to be escaped; so we have to escape both of them -> 4 backslashes.
+        # In the regular expression every backslash, of those 4, needs to be
+        # escaped, leading to 8. This should be without effect on Linux.
+        file = re.sub("\\\\", "\\\\\\\\", file)
+        query = """
         load data local infile '{path}'
         {replace}
         into table `{table}`
@@ -169,9 +173,12 @@ class Connection(SuperConnection):
         escaped by ''
         lines terminated by '\\n'
         ({columns});""".format(
-                path=file.name, replace=replace, table=table, columns=self._sql_cols(df.names)
+                path=file,
+                replace=replace,
+                table=table,
+                columns=self._sql_cols(df.names),
             )
-        )
+        self.bulk_query(query)
 
     def _insert_update(self, df, table, tmp_table):
         self.bulk_query(